requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=1.0.0
joblib>=1.0.0
Flask>=2.0.0
Flask-WTF>=1.0.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.1.0
Werkzeug>=2.0.0
email-validator>=1.1.0 